javamosh: Implement abstract class and interface for tax
calculator (main)

This commit introduces an abstract class  `AbstractTaxCalculator` and an interface
`TaxCalculator` to improve the structure and
reusability of the tax calculation logic.  The
`TaxCalculator2031` class now extends the
abstract class.  The `getTaxableIncome` method
is now implemented in the abstract class and
interface, improving code organization and
reducing redundancy.  A `minimumTax` constant
is added to the `TaxCalculator` interface.